Remove svelte-multiselect
