+++ frontend/package.json
@@ -57,7 +57,6 @@
 		"svelte-check": "^4.0.0",
 		"svelte-floating-ui": "^1.5.8",
 		"svelte-highlight": "^7.6.0",
-		"svelte-multiselect": "^11.0.0-rc.1",
 		"svelte-popperjs": "^1.3.2",
 		"svelte-preprocess": "^6.0.0",
 		"svelte-range-slider-pips": "^2.3.1",
