+++ frontend/src/lib/components/ServiceLogsInner.svelte
@@ -11,12 +11,11 @@
 	import { sendUserToast } from '$lib/toast'
 	import { onDestroy, tick, untrack } from 'svelte'
 	import { Loader2 } from 'lucide-svelte'
-	import { copyToClipboard, truncateRev } from '$lib/utils'
+	import { copyToClipboard, scroll_into_view_if_needed_polyfill, truncateRev } from '$lib/utils'
 	import LogSnippetViewer from './LogSnippetViewer.svelte'
 	import { Button, Drawer, DrawerContent } from './common'
 	import ClipboardCopy from 'lucide-svelte/icons/clipboard-copy'
 	import AnsiUp from 'ansi_up'
-	import { scroll_into_view_if_needed_polyfill } from './multiselect/utils'
 	import SplitPanesOrColumnOnMobile from './splitPanes/SplitPanesOrColumnOnMobile.svelte'
 	import Select from './select/Select.svelte'
 	import { goto } from '$lib/navigation'
